backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
firebase-admin==6.7.0
pydantic==2.11.3
python-dotenv==1.1.0
python-multipart==0.0.20
httpx==0.28.1
nanoid==2.0.0
